+++ packages/meditrak-app-server/src/routes/sync/PushChangesRoute/index.ts
@@ -1,0 +1,1 @@
+export { PushChangesRoute } from './PushChangesRoute';
